feat: lib と mainを分離
