+++ Makefile
@@ -1,0 +1,2 @@
+watch:
+	cargo watch -x check -x test -x run
